+++ CHANGELOG.md
@@ -2,3 +2,4 @@
  - Reworked teleport logic, now it will teleport to the owner more constantly
  - Now tiny ghasts will teleport to the owner if it havent seen the owner for 2400 ticks(2min)
  - Changed fireball logic. The snowball varient will apply freeze effect instead now.
+ - Added tiny ghast spawn egg. It's in the original spawn egg item group, just after ghast spawn egg.